+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `cli/config`: Add support for environment variable `HOME` in path
 - `scripts/install`: Add skeleton for installation script
 - `cli/list`: Add option to show project name header
+- `cli/link`: Add skeleton to link tasks
 
 ### Changed